huge refactoring of how SM properties are handled
the motivation here was to follow the XSMP spec more closely and allow
application's to set custom properties.  the trigger for this was
GNOME's new-ish _GSM_DesktopFile property, which let's an app associate
a .desktop file with itself, so the SM (or other apps) can show a
localized app name, icon, etc.

instead of keeping track of each of the "official" SM properties, we
just have a GTree of everything the client has set.  this makes a lot of
code more generic, and also now makes it easier to make XfsmProperties
look more like a semi-opaque object.  it also removes a lot of annoying
conversion code from XfsmClient, since XfsmProperties uses GValues
internally now.

along with the _GSM_DesktopFile support, xfce4-session now first tries
to use the app's localized name in the splash screen before falling back
to a guess based on SmProgram.