On startup, open a session with ConsoleKit and set the cookie on XDG_SESSION_COOKIE env (as the doc says), close the session on when the xfce4-session shuts down.
Use Upower (devkit-power) for suspend/hibernate, the session will find out the correct bus name:
'org.freedesktop.UPower' or 'org.freedesktop.DeviceKit.Power'

Use ConsoleKit for shutdown/reboot.

Use HAL for suspend/hibernate/shutdown/reboot as a fallback if upower or ConsoleKit are not found.

Use Sudo as a fallback for shutdown/reboot.

Added support for PolicyKit to check authorization of 'org.freedesktop.upower.suspend' or
'org.freedesktop.upower.hibernate' in case we are using upower as for sleep methods.