Reuse existing ConsoleKit sessions (bug #6685).
According to the ConsoleKit documentation, only one session leader may
exist for a session. An application becomes the leader of a session by
calling org.freedesktop.ConsoleKit.Manager.OpenSession() which returns a
session identifier. This identifier is then made available to other
session clients by setting the XDG_SESSION_COOKIE environment variable.

GDM and upstart open ConsoleKit sessions on their own and desktop
environments are supposed to reuse these sessions instead of opening new
ones. So with this commit, xfce4-session checks for XDG_SESSION_COOKIE
and, if it exists, does not try to call OpenSession in order to
overwrite this environment variable.

I tested this with xfce4-session-logout which finally allows me to
reboot and shutdown again.

Also fix the indentation, use spaces instead of tabs etc.