Add support for starting assistive technologies.
Xfce4-session now launches the required at-spi register
applications and takes care of loading the Gtk modules
if the xfconf key /general/StartAt is enabled.

The at-spi applications are launched before the failsafe
applications (panel, wm, desktop) and waits for a maximum
of 2 seconds until the at-bridge is created.

When accessibility is disabled and Gnome services are enabled
we previously started the at-spi desktop files, this is not
done anymore if accessibility is disabled in Xfsm.