Show all applications in autostart (bug #8014).
Show also non-XFCE applications in the autostart list but
do no enable them. If OnlyShowIn does not list XFCE, show
it cursive in the list and set X-XFCE-Autostart-Override=true
when toggled. This way we don't alter the Hidden and OnlyShowIn
keys for other environments.

If X-XFCE-Autostart-Override is true, skip the OnlyShowIn check
and try to start it (unless Hidden is set etcetc).