Restore ssh and gpg agent starting (bug #7018).
Because we cannot start anything before dbus is initialized (which
has been removed from xinit to make startup better with consolekit),
add it in the startup code.

The agent is started and the variables returned by the process
are loaded in the environment before other applications are started.

The daemons are also closed if xfce4-session is terminated.