Fix the shutdown/reboot in xfsm-manager
With the new GDBUS setup, we can't just call gtk_main_quit to exit
since we'll never run the name_lost function which actually calls
shutdown/reboot. This patch changes the gtk_main_quit calls to be
a manager-quit signal which the main.c source will watch and
react to.