Handle screensaver with it's own object (Bug #12603)
Move the screensaver inhibit, heartbeat, and lock code to its own
set of files we can easily share with xfce4-power-manager. Use the
screensaver's dbus API if it supports it for inhibit and lock calls.
Otherwise, use the heartbeat-command and LockCommand from xfpm and
xfsm if available. Finally, for the lock function, fallback to
trying xdg-screensaver, xflock4, and xscreensaver-command in that
order.