Set logout dialog icons from icon_name
Using pixbufs results in unreliable sizing. While the icon may be
present in the icon theme, the pixbuf may be in a different size from
what was defined in the code (32px) and the pixbuf isn't scaled down
(which would anyway look ugly).
So the (correct) alternative approach is to use icon names instead and
check in the icon theme if the icon is present.