Bug #12044: Add a GnuPG 2.1+ mode
GnuPG 2.1+ comes with significantly rewritten agent support.
For example, it no longer supports outputting the GPG_AGENT_INFO
environment variable or the "--write-env-file" command line option.

Because of that we can't track the gpg-agent we started by its PID, we have
to use instead the (upstream-recommended) way to terminate it via a
dedicated gpgconf command.