xflock4: Add debug function
The debug function is controlled by xfce4-session property
/general/LockDebugFile (string) in Xfconf. If it exist and is non-null,
more output will generated in standard error which will be redirected
to the given file. Also standard output will be redirected to that
file. Besides, shell trace output mode will be switched on.

Note that if all file descriptors available for sh are in use, the
script fails and exits with non-zero exit status when trying to find
a free file descriptor. (Tested when using dash as sh.)