Allows update on restart or shutdown when this feature are available.
Offline updates are available since systemd 183 (May 31, 2012), however
it is a feature that of course is only used by Gnome.

See:
 * https://freedesktop.org/wiki/Software/systemd/SystemUpdates/
 * https://github.com/PackageKit/PackageKit/blob/main/docs/offline-updates.txt

The updates are downloaded by PackageKit (`pkcon --only-download update`
or `gnome-software`), and here we only detect that there are updates
available to change the buttons from `Shutdown` and `Restart`, to
`Update and Shutdown`, and `Restart and update` respectively.

Next, we must tell PackageKit that we want to apply the updates when
appropriate.

If PackageKit is not installed or systemd is not used, it does not
change any behavior. On the other hand, if the updates are not
downloaded with any of these applications, the behavior does not change
either.