xflock4: Improve support for lockers (#17)
Run the shell command read from Xfconf property /general/LockCommand
in separate shell, if the property exists and is non-null. That way the
command can not change variables of xflock4 or use its functions. The
property can be a complex shell one-liner that is encapsulated within
xflock4. If you want to use a non-forking locker, you should run it in
background by adding '&' after the command in the line to make it fork
anyway so that xflock4 exits before unlocking. If the one-liner fails,
xflock4 fails, with its exit status. Note that if you put a locker in
background, its exit status will not be received.

xscreensaver-command can take GNU style options so it can be called
from the loop.

'gnome-screensaver-command --lock' returns always 0; tested on
gnome-screensaver 3.6.1; the project is unmaintained, so drop
explicit support for that.

Add explicit support for light-locker, as it is integrated with
xfce4-power-manager. (light-locker pauses audio while locking)

Remove hard-coded fallback lockers because user can specify which
locker to use by LockCommand property.

Xflock4 (!20)