Prevent unintentional recursive calls of xflock4 (#120)
xfce4-session property /general/LockCommand may be set to run xflock4,
so the script becomes recursive. Prevent such recursion by using
a lock directory.

The lock directory is created under $XDG_RUNTIME_DIR, if that variable
is set. Otherwise a warning message is displayed and the lock directory
is created under /tmp. See
https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html
about the former directory. If the latter is preferred, system
administrator should ensure sticky bit is set for /tmp, and remove the
possible stale lock directory on session start, if /tmp is not cleared
otherwise on boot.