Replace XfceRc with GKeyFile for session file parsing
In preparation for a fix requiring delimiter escaping in string lists,
which GKeyFile supports and XfceRc does not. This commit is not supposed
to introduce any change in behavior, except for the few missing sanity
checks added in the process.

Delimiter escaping was not added to XfceRc because it would have
required a complete redesign, which is not justified since GKeyFile
exists and includes all the features of XfceRc.

This commit keeps XfceRc for desktop file parsing where it does not seem
to be a problem, and where it is widely used in the Xfce code base.

Note that this change is backward compatible but not forward compatible,
because GKeyFile adds a delimiter at the end of the list when writing,
unlike XfceRc. And if GKeyFile is able to read correctly a list even if
it does not end with a delimiter, XfceRc interprets the final delimiter
as the presence of an additional empty item.

Related: #1
(cherry picked from commit bda9a260c3a6b901cd9925e0e538fe5b4620c87f)