Replace XfceRc with GKeyFile for session file parsing
In preparation for a fix requiring delimiter escaping in string lists,
which GKeyFile supports and XfceRc does not. This commit is not supposed
to introduce any change in behavior, except for the few missing sanity
checks added in the process.

Delimiter escaping was not added to XfceRc because it would have
required a complete redesign, which is not justified since GKeyFile
exists and includes all the features of XfceRc.

This commit keeps XfceRc for desktop file parsing where it does not seem
to be a problem, and where it is widely used in the Xfce code base.

Related: #1