Add Lock() D-Bus method and make xflock4 a simple wrapper around it
libxfce4ui!97 and 7c846f57 brought behaviors of XfceScreensaver and
xflock4 close enough to take the step, and thus have only one locking
behavior on Xfce, despite several entry points: shell script, D-Bus,
instantiation of XfceScreensaver.

This allows to keep xflock4 which is used in various places in the Xfce
code base and for a long time by users, without needing to make it a
binary instantiating XfceScreensaver.

As for the D-Bus method, it can be used in other contexts, for example
where a proxy on org.xfce.SessionManager is already used for other
things, again avoiding instantiating XfceScreensaver and thus depending
on libxfce4ui.

Fixes: #149
Related: libxfce4ui!97