labwc: Add window snapping range
The default with labwc currently is 1, meaning that if you want to snap
a window to the edge of a monitor, your cursor has to be on the exact
pixel for it to register. This results in a bad experience when using
multiple monitors, as it becomes very difficult to snap windows to the
border between the monitors. The default will be changed to 10 in a
future version of labwc, but it would probably be better to set it
explicitly in the Xfce config to ensure the same behavior regardless of
labwc version.

Related: https://github.com/labwc/labwc/commit/9ca0b111f482f1f8473c6695a953c06464fed6a4
Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>
(cherry picked from commit f1cd45b00506a41061d14e5f67cd6a6f837d5619)